Make sure order of analyses is saved properly. - Fixes https://github.com/jasp-stats/jasp-issues/issues/639 - Also moved all the functions for moving analyses together and into Analyses, Im not sure why part of it was on MainWindow.

+++ JASP-Desktop/results/resultsjsinterface.h
@@ -89,7 +89,7 @@ public slots:
 	void setResultsMetaFromJavascript(QString json);
 	void removeAnalysis(Analysis *analysis);
 	void removeAnalyses();
-	void moveAnalyses(size_t fromId, size_t toId);
+	void moveAnalyses(quint64 fromId, quint64 toId);
 	void setThemeCss(QString themeName);
 
 //end callables
